hooks to tsx
