+++ src/components/postItem/PostItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useDeleteBurn } from '../../hooks/fetchMutations.jsx';
+import { useDeleteBurn } from '../../hooks/fetchMutations';
 import UpdatePostItem from './UpdatePostItem';
 import { useAuth } from '../../context/authContext';
 import '../form/Form.jsx';
