+++ src/hooks/useModal.tsx
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 
+// this is a hook to toggle a modal but it is not being used
+
 const useModal = () => {
   const [isShowing, setIsShowing] = useState(false);
